% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/err.R
\name{err}
\alias{err}
\title{Access the standard error stream}
\usage{
err(...)
}
\arguments{
\item{...}{Character vectors supporting \strong{glue} strings and \strong{cli} inline
styles. This should start with a supported prefix, to map the action to one
of the {rui} verbs. Otherwise, \code{rui::tell()} is used. Supported prefixes are(entitle), i (inform), v (approve), x (disapprove), ~ (begin/proceed), =

(give), * (suggest), ? (ask), ! (alert), w (warn), e (stop), . (display), and
$ (expose). Status bars can be resolved by providing single characters "c",
"v", or "x" for clear, succeed and fail respectively.}
}
\description{
This function provides an alternative to the verb-based API, by mapping the
set of prefixes to the corresponding verbs.
}
